requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
fake-useragent==0.1.11
idna==2.10
lxml==4.6.1
parse==1.18.0
psycopg2==2.8.6
pyee==7.0.4
pyppeteer==0.2.2
pyquery==1.4.1
PyYAML==5.3.1
requests==2.24.0
requests-html==0.10.0
six==1.15.0
soupsieve==2.0.1
tqdm==4.50.2
urllib3==1.25.11
w3lib==1.22.0
websockets==8.1
